requirements.txt
flask
sentence-transformers
supabase
groq
python-dotenv
pypdf
python-telegram-bot
gunicorn